refactor: バレルファイル（index.ts）を削除
バレルファイルによるre-exportをすべて削除し、直接ファイルパスでインポートするようにリファクタリングしました。また、CLAUDE.mdにバレルファイル禁止ルールを追加しました。

- src/components/layout/index.ts を削除
- src/components/auth/index.ts を削除
- src/components/kanban/index.ts を削除
- src/components/repository/index.ts を削除
- src/hooks/index.ts を削除
- src/types/index.ts を削除
- src/stores/index.ts を削除
- src/i18n/index.ts を src/i18n/config.ts にリネーム
- 30+ ファイルのインポートを直接ファイルパスに変更
- CLAUDE.md に「バレルファイル禁止」ルールを追加

型チェック・テスト・ビルドすべて成功
