--- src/components/auth/index.ts
@@ -1,2 +1,0 @@
-export { LoginButton } from "./LoginButton";
-export { UserMenu } from "./UserMenu";
