--- src/components/kanban/index.ts
@@ -1,4 +1,0 @@
-export { KanbanBoard } from "./KanbanBoard";
-export { KanbanColumn } from "./KanbanColumn";
-export { IssueCard } from "./IssueCard";
-export { ColumnSettings } from "./ColumnSettings";
