--- src/components/layout/index.ts
@@ -1,3 +1,0 @@
-export { Header } from "./Header";
-export { Sidebar } from "./Sidebar";
-export { MainLayout } from "./MainLayout";
