--- src/components/repository/index.ts
@@ -1,3 +1,0 @@
-export { RepositoryList } from "./RepositoryList";
-export { RepositoryForm } from "./RepositoryForm";
-export { RepositorySelector } from "./RepositorySelector";
