+++ src/routes/index.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from "@tanstack/react-router";
-import { KanbanBoard } from "../components/kanban";
+import { KanbanBoard } from "../components/kanban/KanbanBoard";
 
 export const Route = createFileRoute("/")({
   component: HomePage,
