--- src/types/index.ts
@@ -1,12 +1,0 @@
-export type { ThemeMode, Locale, Settings } from "./settings";
-export type {
-  Issue,
-  IssueLabel,
-  IssuePriority,
-  KanbanColumn,
-  DragItemType,
-  DragData,
-} from "./kanban";
-export type { GitHubUser, AuthState, AuthStatus } from "./auth";
-export type { Result, OkResult, ErrorResult } from "./result";
-export type { Repository, RepositoryFormData } from "./repository";
